start moving ops out of teletype.c
