+++ simulator/Makefile
@@ -1,8 +1,9 @@
 .PHONY: clean format
 CFLAGS=-std=c99 -g -Wall -DSIM -I. -I../src -I../libavr32/src
 DEPS =
-OBJ = tt.o ../src/teletype.o ../src/state.o \
-	../src/table.o ../src/euclidean/euclidean.o ../src/euclidean/data.o \
+OBJ = tt.o ../src/teletype.o ../src/state.o ../src/table.o \
+	../src/ops/op.o \
+	../src/euclidean/euclidean.o ../src/euclidean/data.o \
 	../libavr32/src/util.o
 
 %.o: %.c $(DEPS)
@@ -16,6 +17,7 @@ clean:
 	rm -f *.o
 	rm -f ../src/*.o
 	rm -f ../src/euclidean/*.o
+	rm -f ../src/ops/*.o
 	rm -f ../libavr32/src/*.o
 
 format:
